Delete DatePicker (and related components) and replace with DatePicker3
