+++ packages/datetime/src/blueprint-datetime.scss
@@ -1,13 +1,11 @@
 // Copyright 2015 Palantir Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
 
-@import "components/date-picker/datepicker";
-@import "components/time-picker/timepicker";
-
-// datetime2
 @import "common/react-day-picker-overrides";
 @import "components/date-input3/dateinput";
+@import "components/date-picker3/datepicker";
 @import "components/date-picker3/date-picker3";
 @import "components/date-picker3/date-picker3-caption";
 @import "components/date-range-picker3/daterangepicker";
 @import "components/date-range-picker3/date-range-picker3";
+@import "components/time-picker/timepicker";
