+++ packages/datetime/src/components/date-picker3/_datepicker.scss
