+++ packages/datetime/src/components/date-range-picker3/dateRangePicker3.tsx
@@ -26,8 +26,8 @@ import { dayPickerClassNameOverrides } from "../../common/classes";
 import { DateRangeSelectionStrategy } from "../../common/dateRangeSelectionStrategy";
 import { combineModifiers, HOVERED_RANGE_MODIFIER } from "../../common/dayPickerModifiers";
 import { MonthAndYear } from "../../common/monthAndYear";
-import { DatePickerUtils } from "../date-picker/datePickerUtils";
 import { DatePicker3Provider } from "../date-picker3/datePicker3Context";
+import { DatePickerUtils } from "../date-picker3/datePickerUtils";
 import { DateFnsLocalizedComponent } from "../dateFnsLocalizedComponent";
 import { DatePickerShortcutMenu, type DateRangeShortcut } from "../shortcuts/shortcuts";
 import { TimePicker } from "../time-picker/timePicker";
