+++ packages/datetime/src/index.md
@@ -37,7 +37,7 @@ Import the package stylesheet in Sass:
 <link href="path/to/node_modules/@blueprintjs/datetime/lib/css/blueprint-datetime.css" rel="stylesheet" />
 ```
 
-@page datepicker
+@page date-picker3
 @page date-input3
 @page date-range-picker3
 @page date-range-input3
